Jg/taskUpdade, atualização importante no packegeJson para vercel (#41)
* feat: Add JWT_SECRET to .env.example and update AuthService

The code changes include adding the JWT_SECRET environment variable to the .env.example file and updating the AuthService to use the JWT_SECRET when creating a token. This change improves the security of the authentication process.

Recent user commits and repository commits indicate a consistent use of the "feat" prefix for feature additions and "refactor" prefix for code refactoring.

* feat: Add AuthModule to app.module.ts

The code changes include adding the AuthModule to the app.module.ts file, enabling authentication functionality in the application. This change enhances the security and user authentication capabilities of the system.

* feat: Add OpcaoModule, ParticipacoesModule, and PerguntasModule to app.module.ts

The code changes include adding the OpcaoModule, ParticipacoesModule, and PerguntasModule to the app.module.ts file. This change enables the functionality related to options, participation, and questions in the application. It enhances the overall functionality and user experience of the system.

Recent user commits and repository commits indicate a consistent use of the "feat" prefix for feature additions and "refactor" prefix for code refactoring.

* feat: Add postinstall script to package.json

The code changes include adding a postinstall script to the package.json file. This script generates the Prisma client and pushes the database changes using the Prisma CLI. This change improves the development workflow by automatically generating the Prisma client and keeping the database schema up to date.

Recent user commits and repository commits indicate a consistent use of the "feat" prefix for feature additions and "refactor" prefix for code refactoring.