refactor: Update PesquisaController to use code instead of id for arquivar endpoint
The code changes in `PesquisaController` modify the `updateArquivar` endpoint to use a `code` parameter instead of an `id` parameter. This change improves the clarity and consistency of the API by using a more descriptive parameter name. The corresponding changes in `PesquisaService` update the database query to use the `code` field instead of the `id` field.

Recent commits:
- refactor: Modify ParticipacaoService to send confirmation email
- refactor: Return message object in PesquisaService
- Merge pull request #82 from juserrrrr/I/Task-teste-unitario-usuariocontroller
- Merge pull request #83 from juserrrrr/GB/Task-Cria-template-de-email-código-de-participação
- Merge pull request #81 from juserrrrr/TS/Auditoria
- chore: Remove console.log statement in upload.service.ts
- Update codigo-voto.html
- Create codigo-voto.html
- fix: renomeia corretamente texto templete